Started on path/route to Castle Village
Started on the Path to Castle Village, added a little skirmish battlefield on the right, plan to add a small camp site on the left side that is off to the side. Going for 3 different routes but only have 2 so far. Working on the third. Main Path is blocked by recently fallen tree.